move connectivity state update test to e2e
